Rename and reorganize HUC components
