+++ Internal/Debug App/Sources/View Controllers/HUC/MerchantHUCCardComponentsViewController.swift
@@ -67,7 +67,7 @@ class MerchantHUCCardComponentsViewController: UIViewController, PrimerHeadlessU
 
         self.cardFormUIManager = try! PrimerHeadlessUniversalCheckout.CardComponentsUIManager()
 
-        var tmpInputElements: [PrimerInputElement] = []
+        var tmpInputElements: [PrimerHeadlessUniversalCheckoutInputElement] = []
         for inputElementType in self.cardFormUIManager!.requiredInputElementTypes {
             let textField = PrimerInputTextField(type: inputElementType, frame: .zero)
             textField.borderStyle = .line
@@ -249,15 +249,15 @@ extension MerchantHUCCardComponentsViewController {
 
 extension MerchantHUCCardComponentsViewController: PrimerInputElementDelegate {
     
-    func inputElementDidFocus(_ sender: PrimerInputElement) {
+    func inputElementDidFocus(_ sender: PrimerHeadlessUniversalCheckoutInputElement) {
 
     }
 
-    func inputElementDidBlur(_ sender: PrimerInputElement) {
+    func inputElementDidBlur(_ sender: PrimerHeadlessUniversalCheckoutInputElement) {
 
     }
 
-    func inputElementValueDidChange(_ sender: PrimerInputElement) {
+    func inputElementValueDidChange(_ sender: PrimerHeadlessUniversalCheckoutInputElement) {
 
     }
 }
