+++ Sources/PrimerSDK/Classes/PCI/Checkout Components/PrimerInputElements.swift
@@ -10,16 +10,17 @@
 import UIKit
 
 public class PrimerInputElementDelegateContainer {
-    var element: PrimerInputElement
+    
+    var element: PrimerHeadlessUniversalCheckoutInputElement
     var delegate: PrimerInputElementDelegate
     
-    init(element: PrimerInputElement, delegate: PrimerInputElementDelegate) {
+    init(element: PrimerHeadlessUniversalCheckoutInputElement, delegate: PrimerInputElementDelegate) {
         self.element = element
         self.delegate = delegate
     }
 }
 
-public class PrimerInputTextField: UITextField, PrimerInputElement {
+public class PrimerInputTextField: UITextField, PrimerHeadlessUniversalCheckoutInputElement {
     public weak var inputElementDelegate: PrimerInputElementDelegate! {
         didSet {
             self.checkoutModulesTextFieldDelegate = PrimerHeadlessUniversalCheckout.Delegate(inputElement: self, inputElementDelegate: inputElementDelegate)
